+++ CHANGELOG.md
@@ -5,7 +5,20 @@ We are currently working on porting this changelog to the specifications in
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## Version 0.7.2 - Unreleased
+## Version 0.7.3 - Unreleased
+
+### Added
+* Add `datetime` attribute to `CocoImage`.
+* Add `ensure_video` method to `CocoDataset`.
+* Add `ensure_track` method to `CocoDataset`.
+* Add `sort_values` method to `ObjectList1D` (and hence vectorized `Images` and `Annots` objects).
+* Add `--inplace` to `kwcoco conform`.
+
+### Fixed
+* kwcoco.Coerce with sqlview now works for .kwcoco.zip files.
+
+
+## Version 0.7.2 - Released 2023-10-04
 
 ### Fixed
 * `clear_annotations` now also clears tracks.